Found case using sleepy store where even though the required number of reads were performed the timeout would abort the entire operation as it still waited on the slow store.
As a result, the code was restructured to wait for responses inline, in the same action so that this case could be properly handled. As a result, the Acknowledge Response actions were removed and thus there wasn't any need for the "event data" to be passed around (as it was only used here), thus simplifying things a bit.